creating a darta.dart file thats have a lot of consts
